pyproject.toml
[project]
name = "prepare-annotations"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "antonkulaga", email = "antonkulaga@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.13.2",
    "eliot>=1.17.5",
    "fsspec>=2025.12.0",
    "huggingface-hub>=1.2.3",
    "polars-bio>=0.19.0",
    "pooch>=1.8.2",
    "prefect>=3.6.7",
    "psutil>=7.1.3",
    "pycomfort>=0.0.18",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "rich>=14.2.0",
    "tenacity>=9.1.2",
    "typer>=0.19.2",
    "vortex-data>=0.56.0",
]

[project.scripts]
prepare-annotations = "prepare_annotations.cli:app"

[build-system]
requires = ["uv_build>=0.9.7,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
